TEST: Optimize e2e watch
